Metaprogramming refactor for default props
Use a decorator to expose a `mergedProps` argument to components.

Next: Do away with decorator approach. It's kind of pointless when I can just call a function to merge the defaultProps. The important insight is I have too keep non-default props separate from merge-able props anyway. Might as well just operate directly on the defaultProps then. Also, there is no need to expose the Immutable object at all. I don't think it costs much.

A lot of files are broken right now, but committing to save the decorator in case I need it.