Lifted padding outside useView. Option to batch animations.  Next: Use SVG.
I tried debugging the sawtooth but I don't think this is the problem after all. The CPU usage is too high, and the sawtooth is periodic with no positive trend. So I think the memory pattern is just inevitable and its the canvas draw calls after all. (Researched setInterval online and tried with no animation, just a timeout. Memory usage I don't think is the problem.)